moved documents related code into the utils file
